PIRANHA-1200 3.0 'Reboot' action does not applied to the instance on the Instance Grid

+++ site/modules/machine/static/js/controllers/machines.js
@@ -433,9 +433,11 @@
                                     'Restart selected instances'
                                 ), function() {
                                     $scope.machines.forEach(function (el) {
-                                        if(el.checked){
-                                            Machine.rebootMachine(id);
-                                            $$track.event('machine', 'reboot');
+                                        if (el.checked) {
+                                            if (el.state != "stopped") {
+                                                Machine.rebootMachine (el.id);
+                                                $$track.event ('machine', 'reboot');
+                                            }
                                             el.checked = false;
                                         }
                                     });
